requirements.txt
charset-normalizer==2.0.11
factory-boy==3.2.1
Faker==12.3.0
idna==3.3
python-dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
six==1.16.0
urllib3==1.26.8